Fix formatting and move to components folder
