+++ packages/frontend/components/MessageList.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Message from './Message'
+
+export default function MessageList({ messages }) {
+  return messages.map((message) => (
+    <Message key={message.id} message={message} />
+  ))
+}
